requirements/rocm-test.txt
# Common dependencies
-r common.txt

# Test infrastructure
tblib==3.1.0
pytest==8.3.5
pytest-asyncio==0.24.0
pytest-timeout==2.3.1
pytest-cov==6.3.0
pytest-forked==1.6.0
pytest-rerunfailures==14.0
pytest-shard==0.1.2

# Async/HTTP dependencies
anyio==4.6.2.post1
    # via httpx, starlette
aiohttp==3.13.3
    # via gpt-oss
httpx==0.27.2
    # HTTP testing

# Audio processing dependencies
audioread==3.0.1
    # via librosa
cffi==1.17.1
    # via soundfile
decorator==5.2.1
    # via librosa
lazy-loader==0.4
    # via librosa
platformdirs==4.3.6
    # via pooch
pooch==1.8.2
    # via librosa
soundfile==0.13.1
    # via librosa
soxr==0.5.0.post1
    # via librosa
librosa==0.10.2.post1

# Retrieval and search
bm25s==0.2.13
    # via mteb
pystemmer==3.0.0
    # via mteb

# Multi-modal processing
blobfile==3.0.0
    # Multi-Modal Models Test
decord==0.6.0
    # video processing, required by entrypoints/openai/test_video.py
rapidfuzz==3.12.1

# OpenAI compatibility and testing
gpt-oss==0.0.8
    # OpenAI compatibility tests
schemathesis==3.39.15
    # OpenAI schema test

# Evaluation and benchmarking
lm-eval[api]>=0.4.9.2
jiwer==4.0.0

# Required for multiprocessed tests that use spawn method, Datasets and Evaluate Test
multiprocess==0.70.16

# Required for v1/metrics/test_engine_logger_apis.py
ray[cgraph,default]>=2.48.0

# Plugins test
terratorch @ git+https://github.com/IBM/terratorch.git@07184fcf91a1324f831ff521dd238d97fe350e3e
torchgeo==0.7.0
    # via terratorch
# MTEB Benchmark Test
mteb==2.1.2

# Utilities
num2words==0.5.14
    # via lm-eval
pqdm==0.2.0
    # via lm-eval

# Required for fastsafetensors test
fastsafetensors @ git+https://github.com/foundation-model-stack/fastsafetensors.git@d6f998a03432b2452f8de2bb5cefb5af9795d459
# Required for suffix decoding test
arctic-inference == 0.1.1
# Required for Nemotron test
open-clip-torch==2.32.0
# Required for isaac Multi-Modal generation test
perceptron==0.1.4
# Required for the multi-modal models test
timm==1.0.17
# Required for plugins test
albumentations==1.4.6
# Pin transformers version
transformers==4.57.3
# Pin HF Hub version
huggingface-hub==0.36.1
